Moved boost:format log proxy to CLoggerBase
