+++ lib/StdInc.h
@@ -2,9 +2,6 @@
 
 #include "../Global.h"
 
-//todo: move to Global.h
-#include <vstd/LogFormat.h>
-
 // This header should be treated as a pre compiled header file(PCH) in the compiler building settings.
 
 // Here you can add specific libraries and macros which are specific to this project.
