+++ lib/VCMI_lib.cbp
@@ -122,7 +122,6 @@
 		</Linker>
 		<Unit filename="../Global.h" />
 		<Unit filename="../include/vstd/CLoggerBase.h" />
-		<Unit filename="../include/vstd/LogFormat.h" />
 		<Unit filename="AI_Base.h" />
 		<Unit filename="BattleAction.cpp" />
 		<Unit filename="BattleAction.h" />
